+++ src/main/java/com/erigitic/config/TEAccount.java
@@ -121,6 +121,13 @@ public class TEAccount implements UniqueAccount {
             accountConfig.getNode(uuid.toString(), currencyName + "-balance").setValue(newBalance.setScale(2, BigDecimal.ROUND_DOWN));
             accountManager.saveAccountConfig();
 
+            // Reset balance to the money cap if it goes over
+            if (totalEconomy.isLoadMoneyCap()) {
+                if (getBalance(currency, contexts).compareTo(totalEconomy.getMoneyCap()) == 1) {
+                    accountConfig.getNode(uuid.toString(), currencyName + "-balance").setValue(totalEconomy.getMoneyCap());
+                }
+            }
+
             transactionResult = new TETransactionResult(this, currency, amount, contexts, ResultType.SUCCESS, TransactionTypes.DEPOSIT);
             totalEconomy.getGame().getEventManager().post(new TEEconomyTransactionEvent(transactionResult));